#30 Rename use case folders
