+++ lib/features/new_book/domain/usecases/store_new_book.dart
