+++ lib/main.dart
@@ -10,11 +10,11 @@ import 'package:storyscape/features/book_storage/data/data_sources/local/book_is
 import 'package:storyscape/features/book_storage/data/data_sources/local/models/local_book_isar_model.dart';
 import 'package:storyscape/features/new_book/data/repositories/book_repository_impl.dart';
 import 'package:storyscape/features/new_book/domain/repositories/book_repository.dart';
-import 'package:storyscape/features/new_book/domain/use_cases/store_new_book.dart';
+import 'package:storyscape/features/new_book/domain/usecases/store_new_book.dart';
 import 'package:storyscape/features/read_book/ui/cubit/book_reader_cubit.dart';
 import 'package:storyscape/features/select_book/data/repositories/available_book_repository_impl.dart';
 import 'package:storyscape/features/select_book/domain/repositories/available_book_repository.dart';
-import 'package:storyscape/features/select_book/domain/use_cases/retrieve_available_books.dart';
+import 'package:storyscape/features/select_book/domain/usecases/retrieve_available_books.dart';
 import 'package:storyscape/features/select_book/ui/cubit/book_selection_cubit.dart';
 import 'package:storyscape/storyscape.dart';
 
@@ -71,8 +71,8 @@ void _setUpBookStorageInjections(GetIt locator) {
       BookIsarDataSourceImpl(
         isar: locator.get<Isar>(),
       ),
-    );
-  locator..registerLazySingleton<LocalBookIsarModelMapper>(LocalBookIsarModelMapperImpl.new);
+    )
+    ..registerLazySingleton<LocalBookIsarModelMapper>(LocalBookIsarModelMapperImpl.new);
 }
 
 void _setUpNewBookInjections(GetIt locator) {
